+++ test/pkgrepl.jl
@@ -54,14 +54,26 @@ end
 
 @testitem "run" begin
     include("setup.jl")
-    CondaPkg.add("python", version="==3.10.2")
     if !isnull
+        CondaPkg.add("python", version="==3.10.2")
         fn = tempname()
+        # run python --version and check the output
         open(fn, "w") do io
             redirect_stdout(io) do
                 CondaPkg.PkgREPL.run(["python", "--version"])
             end
         end
         @test contains(read(fn, String), "3.10.2")
+        # run conda --help and check the output
+        # tests that conda and mamba both run whatever CondaPkg runs
+        for cmd in ["conda", "mamba"]
+            open(fn, "w") do io
+                redirect_stdout(io) do 
+                    CondaPkg.PkgREPL.run(["conda", "--help"])
+                end
+            end
+            @test contains(read(fn, String), "--help")
+            @test contains(read(fn, String), "--version")
+        end
     end
 end