+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+* `pkg> conda run conda ...` now runs whatever conda executable CondaPkg is configured with.
+
 ## 0.2.21 (2023-09-30)
 * Special handling of `python` with `build="**cpython**"`.